Minor refactoring; adjust walking speed and delay buffer
